Move getDefaultOperator to helpers file
